requirements.txt
Django==3.2.17
djangorestframework==3.14.0
drf_yasg==1.21.5
